pyproject.toml
[project]
name = "livraria"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "", email = ""},
]
dependencies = ["django>=5.2.1", "djangorestframework>=3.16.0", "django-cors-headers>=4.7.0", "djangorestframework-simplejwt>=5.5.0", "setuptools>=80.8.0"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
